Update supporting scripts - 0.11.0.

+++ SupportingScripts/Runtime/Shaders/ClippingVolume-MapsSDK.cginc
@@ -5,7 +5,7 @@
 float3 _ClippingVolumePosition;
 float4 _ClippingVolumeNormals[2];
 float3 _ClippingVolumeUp;
-fixed4 _ClippingVolumeColor;
+float4 _ClippingVolumeColor;
 float3 _ClippingVolumeSize;
 float _ClippingVolumeFadeDistance;
